more stable (and insanely asynchronous) room joining
The server no longer makes the assumption that joining
a room is instantaneous, but instead waits for the client
to say it's all ready. This also gives us the opportunity
to animate between rooms, and not ask to be spawned until
the animation is done.

This fix allows us to reintroduce the optimization that
only sends entity updates for entities in the room the
player is in.